// generated/api/models/applicationStatus.ts
/**
 * Generated by orval v7.6.0 🍺
 * Do not edit manually.
 * WolfBankGateway
 * OpenAPI spec version: 1.0
 */

export type ApplicationStatus = (typeof ApplicationStatus)[keyof typeof ApplicationStatus];

// eslint-disable-next-line @typescript-eslint/no-redeclare
export const ApplicationStatus = {
  NUMBER_0: 0,
  NUMBER_1: 1,
  NUMBER_2: 2,
  NUMBER_3: 3,
  NUMBER_4: 4,
  NUMBER_5: 5,
  NUMBER_6: 6
} as const;